add get song by id and fix migration

+++ migrations/1624769836801_create-table-songs.js
@@ -16,7 +16,7 @@ exports.up = (pgm) => {
     },
 
     year: {
-      type: 'NUMERIC',
+      type: 'REAL',
       notNull: true,
     },
 
@@ -31,7 +31,7 @@ exports.up = (pgm) => {
     },
 
     duration: {
-      type: 'NUMERIC',
+      type: 'REAL',
       notNull: true,
     },
